requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
bleach==4.1.0
blessed==1.19.1
cachetools==5.2.0
catboost==1.0.6
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.4
cliff==3.10.1
cloudpickle==2.1.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.4
colorful==0.5.4
colorlog==6.6.0
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.5
docutils==0.17.1
filelock==3.8.0
fonttools==4.34.4
frozenlist==1.3.1
google-api-core==2.8.2
google-auth==2.10.0
googleapis-common-protos==1.56.2
gpustat==1.0.0rc1
graphviz==0.20.1
greenlet==1.1.2
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.39.0
grpcio-tools==1.39.0
idna==3.2
importlib-metadata==4.8.1
joblib==1.1.0
jsonschema==4.9.1
keyring==23.1.0
kiwisolver==1.4.4
llvmlite==0.39.0
m5py==0.3.0
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.2
msgpack==1.0.4
multidict==6.0.2
numba==0.56.0
numpy==1.22.4
nvidia-ml-py==11.495.46
opencensus==0.11.0
opencensus-context==0.1.3
optuna==2.10.1
packaging==21.0
pandas==1.4.3
pbr==5.9.0
Pillow==9.2.0
pkginfo==1.7.1
platformdirs==2.5.2
plotly==5.9.0
plus-ba==0.0.2
prettytable==3.3.0
prometheus-client==0.13.1
protobuf==3.17.3
psutil==5.9.1
py-spy==0.3.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pymongo==4.2.0
pymssql==2.2.5
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.1
scipy==1.9.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smart-open==6.0.0
SQLAlchemy==1.4.39
stevedore==4.0.0
tenacity==8.0.1
threadpoolctl==3.1.0
tqdm==4.62.2
twine==3.4.2
urllib3==1.26.6
virtualenv==20.16.3
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xgboost==1.6.1
yarl==1.8.1
zipp==3.5.0


#--extra-index-url https://${GITLAB_USER}:${GITLAB_TOKEN}@gitlab.brique.kr/api/v4/projects/174/packages/pypi/simple
#qpslib-history-manager==0.1.0
#qpslib-message==0.1.0
#qpslib-model-manager==0.1.2
#qpslib-retrain-manager==

#--extra-index-url https://${GITLAB_USER}:${GITLAB_TOKEN}@gitlab.brique.kr/api/v4/projects/187/packages/pypi/simple
#plus-ba==0.0.3